Fixed Mac build error for #97 & updated to auto-create Julia.app.
The current master (before PR #97) downloads a hand-modified version of
electron v0.35.2, hosted on AWS. This version has been modified so that
when it runs, the process is named Julia, the icons are Julia's, and the
menu bar name is Julia.

Before this commit (at 20dfac1), #97 was failing because it downloaded
stock Electron.app, rather than the hand-modified version.

This commit adds logic to install.jl that programatically applies the
same hand-modifications to the downloaded Electron.app:
 - rename the app to Julia.app
 - update its icons.
 - And update its plist to reflect the new name.
Finally, this commit tells process.jl where to look for the newly
renamed binary. (I chose to also rename the binary just to make the
find/replace in the .plist simpler.)