Put back name "matrix" && older structure.
So the problem was that you can't have both a jobs: section and a matrix: section, since apparently they're aliases for the same thing.

Add back branches:only

After this commit, the only change in this PR is just to move the Documentation "stage" up into the main `matrix:include'.