Prevent possible port collision
While it's a small chance, calling `port()` twice for the debug and non-debug ports may lead to an unlucky draw with both ports come up the same value. By changing the `rand` call to select 2 items from the valid port range instead of calling `rand` twice, we can avoid this possibility entirely. Similarly, I've changed `port` to `ports` to reflect that it returns an array.