Increase the try_connect timeout
A timeout of ~3 seconds is too quick when an X11 window is forwarded over SSH.  Yeah, Electron with a remote X11 window is pretty slow to use, but it's handy at times.  In my case, extending it to ~4 seconds was sufficient.  I've extended it to 5 seconds just to be safe for folks with slower connections.